id : Long -> int , employee_id -> employee_no

+++ src/main/java/com/deligence/deli/domain/ProgressInspection.java
@@ -16,7 +16,7 @@ public class ProgressInspection {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long progress_inspection_no;//진척 검수 일련번호
+    private int progress_inspection_no;//진척 검수 일련번호
 
     private LocalDate progress_inspection_date; //검수일
